Use buffers to make compiler allocations more Rust-friendly
The compiler side is nice but the JS side sees a lot of side-effects.
Data just appears in buffers and it's not obvious how it got there.
I ended up having to put in lots of code comments because otherwise
you'd have no idea what was going on. Can't reason locally, etc.

Code comments are sorta OK I suppose but it just makes me wish it was
simple enough not to need them.